omega-prime-delta/backend/cmd/strategy-engine/requirements.txt
kafka-python==2.0.2
numpy==1.24.3
pandas==2.1.4
torch==2.2.0
